Avoid calls to g_slist_index in webKitWebAudioSrcLoop()
webKitWebAudioSrcLoop() calls g_slist_index for each element
to get its index in a list it's iterating over. g_list_index function
uses a sequential seach to find that element, which is clealy
unecessary.
This patch adds a local variable to store the current index and use it
instead of calling g_slist_index funtion.